Continue set up, refactor package, create config directory for configuration files
