--- src/main/java/com/example/securingweb/blahhh.java
@@ -1,43 +1,0 @@
-package com.example.securingweb;
-
-public class blahhh {
-
-
-    public boolean testing(){
-        return true;
-    }
-
-
-    public void pleaseWork(){
-        System.out.println("if you see this means it works");
-        System.out.println("fuck you");
-        System.out.println("fuck 2");
-        System.out.println("mother fucker");
-        System.out.println("mother fucker 2");
-
-    }
-    //this is to test
-
-
-    public static void main(String[] args) {
-        System.out.println("hello");
-
-    }
-
-
-    public int sum(){
-        return 1 + 2;
-    }
-
-
-
-
-    public int testing1(){
-        return 1;
-    }
-
-    public String test(){
-        return "yo";
-    }
-
-}
