+++ src/main/java/io/muzoo/scc/project/backend/SecuringWebApplication.java
@@ -1,4 +1,4 @@
-package com.example.securingweb;
+package io.muzoo.scc.project.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
