+++ src/test/java/io/muzoo/scc/project/backend/SecuringWebApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.securingweb;
+package io.muzoo.scc.project.backend;
 
 import org.junit.jupiter.api.Test;
